chore: separate test
