+++ .github/workflows/unit.yml
@@ -25,26 +25,21 @@ jobs:
       - name: checkout
         uses: actions/checkout@master
 
-      - name: Get yarn cache directory path
-        id: yarn-cache-dir-path
-        run: echo "::set-output name=dir::$(yarn cache dir)"
-
-      - uses: actions/cache@v2
-        id: yarn-cache # check for `cache-hit` (`steps.yarn-cache.outputs.cache-hit != 'true'`)
+      - name: cache node_modules
+        uses: actions/cache@v1
         with:
-          path: ${{ steps.yarn-cache-dir-path.outputs.dir }}
-          key: ${{ runner.os }}-yarn-${{ hashFiles('**/yarn.lock') }}
-          restore-keys: |
-            ${{ runner.os }}-yarn-
+          path: node_modules
+          key: ${{ matrix.os }}-node-v${{ matrix.node }}-deps-${{ hashFiles(format('{0}{1}', github.workspace, '/yarn.lock')) }}
 
       - name: Install dependencies
+        if: steps.cache.outputs.cache-hit != 'true' 
         run: yarn
 
       - name: Lint
         run: yarn lint
 
       - name: Test
-        run: yarn jest -i
+        run: yarn test:unit
 
       - name: Coverage
         uses: codecov/codecov-action@v1
