+++ package.json
@@ -18,7 +18,9 @@
     "dev": "yarn nuxt-ts playground",
     "lint": "eslint --ext .ts --ext .vue .",
     "release": "yarn test && standard-version && yarn build && git push --follow-tags && npm publish",
-    "test": "yarn lint && yarn jest"
+    "test": "yarn lint && yarn test:unit && yarn test:e2e",
+    "test:e2e": "jest test/e2e --forceExit",
+    "test:unit": "jest test/unit --forceExit"
   },
   "dependencies": {
     "hasha": "^5.2.1",
