+++ test/unit/providers.test.ts
@@ -1,8 +1,8 @@
 import fs from 'fs-extra'
-import local from '../src/providers/local'
-import cloudinary from '../src/providers/cloudinary'
-import twicpics from '../src/providers/twicpics'
-import fastly from '../src/providers/fastly'
+import local from '~/src/providers/local'
+import cloudinary from '~/src/providers/cloudinary'
+import twicpics from '~/src/providers/twicpics'
+import fastly from '~/src/providers/fastly'
 
 const images = [
   {
